requirements.txt
matplotlib==3.10.3
numpy==2.3.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
PyYAML==6.0.2
sympy==1.13.1
